change sum.py
